requirement.txt
click==8.1.3
cmake==3.26.3
customtkinter==5.1.3
darkdetect==0.8.0
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
Pillow==9.5.0
